aquasense_backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML
google-generativeai==0.3.2
tensorflow-macos==2.16.2  # For Apple Silicon/Intel Mac (latest available)
pillow==10.2.0
numpy==1.26.3

# Text-to-Speech
gTTS==2.5.0

# WebSocket Support
websockets==12.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
